LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting-api/reporting-api-report-to-only-sends-reports-to-first-endpoint.https.sub-expected.txt



PASS Test that image does not load
PASS Event is fired
FAIL Violation report status OK. assert_equals: CSP report sent, but not expecting one. expected 0 but got 2
